+++ CHANGELOG.md
@@ -65,3 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - **Category:** Only show categories from database.
+
+### Fixed
+
+- **Monthly Income:** Fixed monthly income calculation.